Refactored clever auth into its own module
