+++ src/api/Auth/index.ts
@@ -1,4 +1,3 @@
-export * from './cleverAuth';
 export * from './getMe';
 export * from './login';
 export * from './reset';
