+++ src/api/index.ts
@@ -1,4 +1,5 @@
 export * as Auth from './Auth';
+export * as Clever from './Clever';
 export * as PastStreams from './PastStreams';
 export * as Prompts from './Prompts';
 export * as Submissions from './Submissions';
